Changed the type of ids to long

+++ Mastonet/Entities/Status.cs
@@ -11,7 +11,7 @@ namespace Mastonet.Entities
         /// The ID of the status
         /// </summary>
         [JsonProperty("id")]
-        public int Id { get; set; }
+        public long Id { get; set; }
 
         /// <summary>
         /// A Fediverse-unique resource ID
@@ -35,13 +35,13 @@ namespace Mastonet.Entities
         /// null or the ID of the status it replies to
         /// </summary>
         [JsonProperty("in_reply_to_id")]
-        public int? InReplyToId { get; set; }
+        public long? InReplyToId { get; set; }
 
         /// <summary>
         /// null or the ID of the account it replies to
         /// </summary>
         [JsonProperty("in_reply_to_account_id")]
-        public int? InReplyToAccountId { get; set; }
+        public long? InReplyToAccountId { get; set; }
 
         /// <summary>
         /// null or the reblogged Status
